% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{ClassTopics_results}
\alias{ClassTopics_results}
\title{Extract and Interpret Supervised Topic Modeling Results}
\usage{
ClassTopics_results(
  fit,
  true_response,
  vars_names = NULL,
  top_vars = 10,
  credible_interval = 0.95
)
}
\arguments{
\item{fit}{Stan fit object from \code{ClassTopics}}

\item{true_response}{Character or factor vector on the class labels}

\item{vars_names}{Character vector of variable names (optional)}

\item{top_vars}{Integer, number of top variables to extract per topic}

\item{credible_interval}{Numeric, credible interval width (default: \code{0.95})}
}
\value{
List containing all results and interpretations
}
\description{
Extract and Interpret Supervised Topic Modeling Results
}
\examples{

## Generate synthetic dataset

df <- generate_CTdata()

# Remove zero-count columns
zero_cols <- colSums(df@X) == 0
X <- df@X[, !zero_cols, drop = FALSE]

# Model fitting
\donttest{
fit <- ClassTopics(counts = X, response = df@Y)

# Posterior results extraction
results <- ClassTopics_results(fit = fit, true_response = df@Y)

}

}
