% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CTparameter-class}
\alias{CTparameter-class}
\alias{paramMean}
\alias{paramMean,CTparameter-method}
\alias{paramCI}
\alias{paramCI,CTparameter-method}
\alias{show,CTparameter-method}
\title{Class CTparameter}
\usage{
paramMean(x)

\S4method{paramMean}{CTparameter}(x)

paramCI(x)

\S4method{paramCI}{CTparameter}(x)

\S4method{show}{CTparameter}(object)
}
\arguments{
\item{x}{An CTparameter object.}

\item{object}{An CTparameter object.}
}
\description{
An S4 class representing a single model parameter (e.g. theta, beta, or
eta) summarized as a posterior mean matrix together with lower and upper
credible interval bounds. All three slots are always 2-dimensional and
share identical dimensions.
}
\section{Slots}{

\describe{
\item{\code{mean}}{Numeric matrix of posterior means.}

\item{\code{lower}}{Numeric matrix of lower credible interval bounds.}

\item{\code{upper}}{Numeric matrix of upper credible interval bounds.}
}}

