% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CTparameter}
\alias{CTparameter}
\title{Constructor for CTparameter}
\usage{
CTparameter(mean, lower, upper)
}
\arguments{
\item{mean}{Numeric matrix of posterior means.}

\item{lower}{Numeric matrix of lower credible interval bounds.}

\item{upper}{Numeric matrix of upper credible interval bounds.}
}
\value{
An object of class \code{CTparameter}.
}
\description{
Constructor for CTparameter
}
