% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CTprediction-class}
\alias{CTprediction-class}
\alias{predictedClass}
\alias{predictedClass,CTprediction-method}
\alias{predictedProbs}
\alias{predictedProbs,CTprediction-method}
\alias{show,CTprediction-method}
\title{Class CTprediction}
\usage{
predictedClass(x)

\S4method{predictedClass}{CTprediction}(x)

predictedProbs(x)

\S4method{predictedProbs}{CTprediction}(x)

\S4method{show}{CTprediction}(object)
}
\arguments{
\item{x}{A CTprediction object.}

\item{object}{A CTprediction object.}
}
\description{
An S4 class representing the output of \code{\link{predict_ClassTopics_stan}}:
predicted class labels and predicted class probabilities for new
(out-of-sample) observations, given a fitted model's topic-gene loadings
and regression coefficients.
}
\section{Slots}{

\describe{
\item{\code{predicted_class}}{Character vector of predicted category labels
(one per observation).}

\item{\code{predicted_probs}}{Numeric matrix of predicted class probabilities
(observations x categories).}
}}

