% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CTprediction}
\alias{CTprediction}
\title{Constructor for CTprediction}
\usage{
CTprediction(predicted_class, predicted_probs)
}
\arguments{
\item{predicted_class}{Character vector of predicted category labels.}

\item{predicted_probs}{Numeric matrix of predicted class probabilities.}
}
\value{
An object of class \code{CTprediction}.
}
\description{
Constructor for CTprediction
}
