% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics-simulation.R
\docType{class}
\name{CTsynthetic-class}
\alias{CTsynthetic-class}
\title{Class CTsynthetic}
\description{
An S4 class representing a count dataset simulation following a supervised
non-negative matrix factorization (NMF) approach.
}
\section{Slots}{

\describe{
\item{\code{X}}{Integer matrix of data counts.}

\item{\code{H}}{Numeric matrix of observation-topic loadings.}

\item{\code{W}}{Numeric matrix of topic-variable loadings.}

\item{\code{Y}}{Factor of observed labels}
}}

