% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CTtrainpred-class}
\alias{CTtrainpred-class}
\title{Class CTtrainpred}
\description{
An S4 class representing the summary of the predictive performance of
the fitted multiclass topic model (without cross-validation) on the
respective training data.
}
\section{Slots}{

\describe{
\item{\code{predicted_categories}}{Factor vector of posterior predicted categories.}

\item{\code{true_categories}}{Factor vector of true categories.}

\item{\code{prediction_probabilities}}{Numeric matrix whose rows are the posterior
softmax vectors for each observation.}

\item{\code{overall_accuracy}}{Numeric scalar, accuracy of the model.}

\item{\code{category_accuracy}}{Numeric vector, accuracy of the model stratified by
class.}

\item{\code{confusion_matrix}}{Integer matrix, confusion matrix of the counts at
each \code{predicted_categories}/\code{true_categories} combination.}
}}

