% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CTtrainpred}
\alias{CTtrainpred}
\title{Constructor for CTtrainpred}
\usage{
CTtrainpred(
  predicted_categories,
  true_categories,
  prediction_probabilities,
  overall_accuracy,
  category_accuracy,
  confusion_matrix
)
}
\arguments{
\item{predicted_categories}{Factor vector of posterior predicted categories.}

\item{true_categories}{Factor vector of true categories.}

\item{prediction_probabilities}{Numeric matrix whose rows are the posterior
softmax vectors for each observation.}

\item{overall_accuracy}{Numeric scalar, accuracy of the model.}

\item{category_accuracy}{Numeric vector, accuracy of the model stratified by
class.}

\item{confusion_matrix}{Integer matrix, confusion matrix of the counts at
each \code{predicted_categories}/\code{true_categories} combination.}
}
\value{
An object of class \code{CTtrainpred}.
}
\description{
Constructor for CTtrainpred
}
