% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{.build_cv_list}
\alias{.build_cv_list}
\title{Secondary function that prepares a list to accommodate the results from
each fold in cross-validation}
\usage{
.build_cv_list(D, response_levels, k_folds)
}
\arguments{
\item{D}{number of observations in the full dataset}

\item{response_levels}{character, class labels}

\item{k_folds}{number of folds for cross-validation}
}
\value{
list with elements \code{all_test_predictions}, \code{all_test_pred_probs},
\code{fold_train_accuracies}, \code{fold_test_accuracies},
\code{fold_train_class_accuracies} and \code{fold_test_class_accuracies}
}
\description{
Secondary function that prepares a list to accommodate the results from
each fold in cross-validation
}
