% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{.create_stratified_folds}
\alias{.create_stratified_folds}
\title{Create Stratified Folds}
\usage{
.create_stratified_folds(response, k_folds)
}
\arguments{
\item{response}{Factor vector of class labels to stratify by.}

\item{k_folds}{Integer, number of folds to create.}
}
\value{
A list of length \code{k_folds}, each element an integer vector of
row indices assigned to that fold's test set.
}
\description{
Create Stratified Folds
}
\keyword{internal}
