% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{.cv_ClassTopics_overall}
\alias{.cv_ClassTopics_overall}
\title{K-Fold Cross-Validation for ClassTopics (For unbiased estimation)}
\usage{
.cv_ClassTopics_overall(
  gene_counts,
  response,
  folds,
  pred_n_acc,
  final_model = NULL,
  ...
)
}
\arguments{
\item{gene_counts}{gene count matrix}

\item{response}{character or factor vector on the class labels}

\item{folds}{list whose elements are train-test partitions (as produced
by \link{.create_stratified_folds})}

\item{pred_n_acc}{list containing the results of the previously obtained
classification results with .cv_ClassTopics_fold_by_fold}

\item{final_model}{argument that may receive the estimated full model
obtained with ClassTopics. If \code{NULL} (default), said model is
fitted internally}

\item{...}{Additional arguments passed to ClassTopics and predict_ClassTopics
(both _stan and _EM)}
}
\value{
List with summarized results of the topic model and cross-validated
classification results
}
\description{
K-Fold Cross-Validation for ClassTopics (For unbiased estimation)
}
\details{
This performs k-fold cross-validation:
\itemize{
\item Each patient appears in test set exactly ONCE
\item Training sets are DISJOINT (no overlap)
\item Accuracy estimate is UNBIASED
}

Returns:
\itemize{
\item \code{cv_accuracy}: Unbiased estimate of model performance
\item \code{final_model}: Model trained on ALL data (for interpretation)
}
}
\keyword{internal}
