% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\name{.stan_model_names}
\alias{.stan_model_names}
\title{Names of the Stan models shipped with ClassTopics}
\usage{
.stan_model_names
}
\description{
Names of the Stan models shipped with ClassTopics
}
\keyword{internal}
