% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\name{.stan_source_path}
\alias{.stan_source_path}
\title{Locate the installed .stan source file for a given model}
\usage{
.stan_source_path(model_name)
}
\arguments{
\item{model_name}{One of "model_betadir", "model_pureNMF", "model_test"}
}
\value{
Absolute path to the installed .stan file
}
\description{
Locate the installed .stan source file for a given model
}
\keyword{internal}
