% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics-simulation.R
\name{generate_CTdata}
\alias{generate_CTdata}
\title{Generates a class-balanced synthetic count dataset following a supervised
Poisson non-negative matrix factorization approach}
\usage{
generate_CTdata(
  D_per_class = 50,
  V = 100,
  eta = NULL,
  alpha_w = 0.05,
  lambda = 1e+06,
  class_spread = 2,
  seed = 123
)
}
\arguments{
\item{D_per_class}{integer, number of observations to draw per class
(default = 50)}

\item{V}{integer, number of synthetic variables (default = 100)}

\item{eta}{numeric or matrix with the topic-response coefficients.
If \code{NULL} (default), it gets defined as
\code{eta <- matrix(rnorm(15), 3, 5)}}

\item{alpha_w}{numeric, controls data sparsity (default = 0.05)}

\item{lambda}{numeric, corresponds to the magnitude of the dataset
(default = 1e6)}

\item{class_spread}{numeric that controls level of class separation
(small \code{class_spread} favors class overlapping,
large \code{class_spread} favors class separation).
Default = 2}

\item{seed}{numeric seed for reproducibility (default = 123)}
}
\value{
an object of class CTsynthetic
}
\description{
Generates a class-balanced synthetic count dataset following a supervised
Poisson non-negative matrix factorization approach
}
\details{
When 'eta' is given as a 'matrix' object, its rows correspond
to classes and columns to topics. If 'eta' is a 'numeric' array,
the function assumes the symmetric binary class construction and
executes \code{eta <- rbind(eta, -eta)} before simulation so that
softmax is applied smoothly.
}
\examples{

## Generate with the default parameters

df <- generate_CTdata()

## Specific multinomial case with 3 classes

eta <- matrix(c( 0.4,  0.4,  0.0,  2.0, -1.1, -0.7,
                -0.1,  0.5, -0.3, -1.3,  2.2, -0.1,
                -0.3, -0.9,  0.3, -0.7, -1.1,  0.8),
                nrow = 3, byrow = TRUE)

D_per_class <- 200
V <- 105
eta <- eta
alpha_w <- 0.05
lambda <- 1e6
class_spread <- 2
seed <- 123

df <- generate_CTdata(D_per_class = D_per_class,
                      V = V,
                      eta = eta,
                      alpha_w = alpha_w,
                      lambda = lambda, 
                      class_spread = class_spread,
                      seed = seed)
                      
## Symmetric binary case, reusing all parameters except for 'eta'

eta <- c(-1.1, 0.0, 0.5, -0.9, 0.8)

# Also an option that produces the same result:
# eta <- matrix(c(-1.1, 0.0, 0.5, -0.9, 0.8), nrow = 1)
                
df <- generate_CTdata(D_per_class = D_per_class,
                      V = V,
                      eta = eta,
                      alpha_w = alpha_w,
                      lambda = lambda, 
                      class_spread = class_spread,
                      seed = seed)
                      
## Asymmetric binary case, reusing all parameters except for 'eta'

eta <- matrix(c( 0.4,  0.4,  0.0,  2.0, -1.1, -0.7,
                -0.3, -0.9,  0.3, -0.7, -1.1,  0.8),
                nrow = 2, byrow = TRUE)
                
df <- generate_CTdata(D_per_class = D_per_class,
                      V = V,
                      eta = eta,
                      alpha_w = alpha_w,
                      lambda = lambda, 
                      class_spread = class_spread,
                      seed = seed)

}
