% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{plot_top_genes}
\alias{plot_top_genes}
\title{Plot Top Genes per Topic}
\usage{
plot_top_genes(results, n_genes = 5)
}
\arguments{
\item{results}{Results from \code{\link[=ClassTopics_results]{ClassTopics_results()}} or \code{\link[=cv_ClassTopics_results]{cv_ClassTopics_results()}}}

\item{n_genes}{Integer, number of top genes to show per topic}
}
\value{
A \code{ggplot} object showing the top genes for each topic
}
\description{
Plot Top Genes per Topic
}
