% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{plot_topic_correlations}
\alias{plot_topic_correlations}
\title{Plot Topic Correlations}
\usage{
plot_topic_correlations(results)
}
\arguments{
\item{results}{Results from \code{\link[=ClassTopics_results]{ClassTopics_results()}} or \code{\link[=cv_ClassTopics_results]{cv_ClassTopics_results()}}}
}
\value{
A \code{ggplot} object showing a heatmap of topic correlations
}
\description{
Plot Topic Correlations
}
\examples{

df <- generate_CTdata()

# Remove zero-count columns
zero_cols <- colSums(df@X) == 0
X <- df@X[, !zero_cols, drop = FALSE]

# Model fitting
\donttest{
fit <- ClassTopics(counts = X, response = df@Y)

# Posterior results extraction
results <- ClassTopics_results(fit = fit, true_response = df@Y)

# Plot topic correlations
plot_topic_correlations(results)
}

}
