% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{plot_top_vars}
\alias{plot_top_vars}
\title{Plot Top Variables per Topic}
\usage{
plot_top_vars(results, n_vars = 5)
}
\arguments{
\item{results}{Results from \code{\link[=ClassTopics_results]{ClassTopics_results()}} or \code{\link[=cv_ClassTopics_results]{cv_ClassTopics_results()}}}

\item{n_vars}{Integer, number of top vars to show per topic}
}
\value{
A \code{ggplot} object showing the top vars for each topic
}
\description{
Plot Top Variables per Topic
}
