% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{plot_topic_response_heatmap}
\alias{plot_topic_response_heatmap}
\title{Plot Topic-Response Associations}
\usage{
plot_topic_response_heatmap(results, significance_only = FALSE)
}
\arguments{
\item{results}{Results from \code{\link[=ClassTopics_results]{ClassTopics_results()}} or \code{\link[=cv_ClassTopics_results]{cv_ClassTopics_results()}}}

\item{significance_only}{Logical, show only significant associations}
}
\value{
A \code{ggplot} object showing a heatmap of topic-response coefficients
}
\description{
Plot Topic-Response Associations
}
