% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{predict_ClassTopics_EM}
\alias{predict_ClassTopics_EM}
\title{Predict Mixed Memberships and Response for New Patients Given Learned Topics via EM only}
\usage{
predict_ClassTopics_EM(
  counts_test,
  W,
  eta,
  response,
  shape = 1,
  n_iter_EM = 4,
  verbose = TRUE
)
}
\arguments{
\item{counts_test}{Matrix of counts}

\item{W}{Learned topic-variable loadings \code{K, V}}

\item{eta}{Learned regression coefficients \code{C, K}}

\item{response}{Character vector of response categories}

\item{shape}{scalar prior shape of H (default = 1)}

\item{n_iter_EM}{Number of required EM iterations}

\item{verbose}{logical that controls whether EM likelihood is printed
(default = \code{TRUE})}
}
\value{
List with predicted classes and probabilities
}
\description{
Predict Mixed Memberships and Response for New Patients Given Learned Topics via EM only
}
