% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTopics.R
\name{predict_ClassTopics_stan}
\alias{predict_ClassTopics_stan}
\title{Predict Mixed Memberships and Response for New Observations Given Learned Topics via Stan}
\usage{
predict_ClassTopics_stan(
  counts_test,
  W,
  eta,
  mu_target,
  response,
  cores = 3,
  chains = 3,
  control = list(adapt_delta = 0.9, max_treedepth = 15),
  seed = 123,
  iter_warmup = 1000,
  iter_sampling = 1000,
  n_iter_EM = 4,
  verbose = TRUE
)
}
\arguments{
\item{counts_test}{Matrix of counts}

\item{W}{Learned topic-variable loadings \code{K, V}}

\item{eta}{Learned regression coefficients \code{C, K}}

\item{mu_target}{reciprocal of the rate parameter}

\item{response}{Character vector of response categories}

\item{cores}{number of CPU cores to use (default = \code{3})}

\item{chains}{number of chains to run (default = \code{3})}

\item{control}{control parameters for the Stan algorithm}

\item{seed}{numeric seed for reproducibility (default = \code{123})}

\item{iter_warmup}{number of warm-up iterations for each chain
(default = \code{1000})}

\item{iter_sampling}{number of sampling iterations for each chain
(default = \code{1000})}

\item{n_iter_EM}{Number of required EM iterations}

\item{verbose}{logical that controls whether EM likelihood is printed
(default = \code{TRUE})}
}
\value{
List with predicted classes and probabilities
}
\description{
Predict Mixed Memberships and Response for New Observations Given Learned Topics via Stan
}
